Limit Coupling Fully Implemented
When gearing axes in the MCPower FB will now incorporate the limits of the slave axes coupled to it. This pays attention to the gearratio implemented such that if a slave is coupled in with a +ve ratio then foward limit applies to forward limit (and with backward) but if the ratio is -ve then the foward limit will apply to the backward enable. This has been implemented for MCGearIn and MCGearInMultiMaster.

The multimaster code is a little long and repetetive and would benefit from the master axis DUT that Dom worked on in his branch.