Gearing Axis structure altered
Previously the AxisControl struct would contain the entire Axis_Ref for the master axes , these have now changed to UINT meaning each axisControl struct only contains a single axis_ref for the axis it is linked to, this makes NC linking much easier and cleaner. The axisControl array has also moved in to the global GVL so that within the AxDriveFB the gear in functions can access the necessary Axis_Ref structures for the masters. This also removes the reliance on axes needing to be linked to the array index that matches the NC axisID value meaning the operation and linking of limits when gearing should be much more robust.
All HMI elements have been updated to reference GVL.axes rather than MAIN.axes and a nice benefit is the gearing tab can now correctly show the Master axes