No more n^2
Master/slave relationship for axes has now been altered. Previously there was a histogram arrangement where each master could see slaves, this has changed so that slaves hold master information in an array of size 5.
On each cycle an axis will have to check this array size 5 for every axis in the system to see if it is listed as a master in any of them.

The multimaster axes have also been moved in to an array of a new DUT that contains an index and ratio so simplify limits code further when gearing in.